[FAB-18268] Modify setPolicies() to replace all policies in an existing configuration with the given ones
Currently, setPolicies() upserts policies in an existing configuration with
the given ones. This speficifation makes it difficult to remove multiple
policies that are not needed.

In preparation for supporting SetPolicies() for groups and orgs,
this patch modifies setPolicies() to clear all existing policies in an
existing configuration and then replace them with the given policies.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>